Move styles to common folder
